refactor: dockerfiles hierarchy and centos8-minimal checksum
labels at the end of file as one block since some of them change
frequently. Created label is redundant since it is already in the
metadata. Centos8-minimal build now checks checksum of pckg file list
and saves it as a variable. That way any change will trigger the build.